+++ frontend/svelte/src/lib/services/proposals.services.ts
